[merge][strongdc-sqlite] r14309 -> r14749:
changelog:
===============

Revision: 14749
Author: pavel.pimenov@gmail.com
Date: 26 июля 2013 г. 21:09:45
Message:
Merged revision(s) 14747 from trunk:
# (for merge) Исправлен баг в static uint32_t toUInt32(const char* c)
  после хоть одного переполнения глобальный фгаг оставался в errno == ERANGE
  и это приводило к двойному преобразованию - сначала uint32_t ret = atoi(c);
  а потом return (uint32_t)_atoi64(c);
........

----

Revision: 14741
Author: pavel.pimenov@gmail.com
Date: 26 июля 2013 г. 7:18:10
Message:
# [strongdc-sqlite] BOOST_USE_WINDOWS_H
----

Revision: 14740
Author: pavel.pimenov@gmail.com
Date: 26 июля 2013 г. 7:17:39
Message:
# [strongdc-sqlite] Вернул работу под XP
----

Revision: 14736
Author: pavel.pimenov@gmail.com
Date: 25 июля 2013 г. 23:33:58
Message:
# [strongdc-sqlite] Отключил VLD в релизе
#ifdef _DEBUG
#define _USE_FLYLINKDC_VLD
#endif

----

Revision: 14735
Author: pavel.pimenov@gmail.com
Date: 25 июля 2013 г. 22:46:19
Message:
* Updating the database ISPs
   http://www.assembla.com/spaces/customlocations-greylink
   Build on the base 1264, created 2013-07-17 10:19:21 +0700
   Total in database:
    - ISP - 1399
    - ISPs logo - 1408
    - Address range - 5940
   Archive CustomLocations.7z created by NightOrion

----

Revision: 14734
Author: pavel.pimenov@gmail.com
Date: 25 июля 2013 г. 22:45:45
Message:
# [strongdc-sqlite] Astyle
----

Revision: 14733
Author: pavel.pimenov@gmail.com
Date: 25 июля 2013 г. 22:45:20
Message:
# [strongdc-sqlite] Исправлен AdcHub::handle
  Убрал лишний поиск в массиве c.getParam("NI", 0, l_nick)
----

Revision: 14713
Author: pavel.pimenov
Date: 25 июля 2013 г. 10:39:53
Message:
* [strongdc-sqlite] [merge]
* (Closed issue 1101) http://code.google.com/p/flylinkdc/issues/detail?id=1101
  При открытии файл-листов где имя файла пустое не кидаем исключение Directory missing name attribute
  а генерируем имя файла фиктивное l_file_name = "empty_file_name_" + Util::toString(++m_empty_file_name_counter);
* (Closed  issue 1107 ) http://code.google.com/p/flylinkdc/issues/detail?id=1107
  Исправлено падение приложения при закрытии если в этот момент идет прасинг файл-листа.
* [r5xx][merge]
  - В файл листах для каталога рассчитываем min-max битрейт
  - Оптимизировано открытие файл-листов оригинальных клиентов (Если на первом файле в листе нет тэга TS, то на следующих файлах не пытаемся искать медиа-тэги их там не будет)
  - QueueManager:
       - Выполняем вызов getDownloadConnection только на первом файле первого каталога при скачке каталога с одного юзера.
       - Исправлено зависание при скачивании каталогов с кол-вом файлов > 10-100 тыс.... + Может тормозить когда качается много каталогов.
* HAVE_NATUPNP_H включил для client.lib
----

Revision: 14686
Author: pavel.pimenov@gmail.com
Date: 23 июля 2013 г. 22:53:00
Message:
# [strongdc-sqlite][merge]
----

Revision: 14633
Author: pavel.pimenov@gmail.com
Date: 22 июля 2013 г. 7:05:45
Message:
# [merge]
----

Revision: 14480
Author: pavel.pimenov
Date: 11 июля 2013 г. 9:53:14
Message:
# [merge]
  http://code.google.com/p/flylinkdc/issues/detail?id=1077
  Исправлена трансляция ошибок от wininet. теперь в журналы выводится не только
  код, но и описание ошибки (for merge)
----

Revision: 14442
Author: pavel.pimenov
Date: 9 июля 2013 г. 8:51:40
Message:
# [merge] Инитим WINDOWPLACEMENT
----

Revision: 14441
Author: pavel.pimenov
Date: 9 июля 2013 г. 8:37:21
Message:
# Поднимаем раньше флаг закрытия главной формы чтобы в холостую не тикали секунды: MainFrame::on(TimerManagerListener::Second
  TODO - merge https://bitbucket.org/3F/flightsdc/commits/bd42ea5216d7395f3221549aa00d4d83613c2345
----

Revision: 14376
Author: pavel.pimenov@gmail.com
Date: 4 июля 2013 г. 18:41:12
Message:
# [merge] В void QueueManager::add убрал создание лишнего временно объекта SearchManager::getInstance()->search(TTHValue(root).toBase32()
----

+++ FlyFeatures/flyServer.cpp
@@ -527,8 +527,8 @@ string CFlyServerJSON::connectFlyServer(const CFlyServerKeyArray& p_fileInfoArra
 	Json::Value  l_root;   	
 	Json::Value& l_arrays = l_root["array"];
 	l_root["header"]["ID"] = CFlyServerJSON::g_fly_server_id;
-	int  l_count = 0;
-	int  l_count_only_ext_info = 0;
+	size_t  l_count = 0;
+	size_t  l_count_only_ext_info = 0;
 	bool is_all_file_only_ext_info = false;
 	if( p_is_fly_set_query == false) // Для get-запроса все записи хотят только расширенную инфу?
 	{
